fix: import cycle in conversion refactor (#315)
